@@ -5,36 +5,55 @@
   "module": "build/dist/index.js",
   "scripts": {
     "clean": "npx shx rm -rf build",
-    "build": "rollup -c",
-    "watch": "rollup -c -w",
+    "build": "npm run clean && NODE_ENV=production rollup -c",
+    "watch": "rollup -c -w --no-watch.clearScreen",
+    "develop":"npm run clean && run-p watch proxy",
     "postinstall": "check-plugin && check-peer-dependencies || true",
     "proxy": "dev-proxy",
-    "modules": "yarn clean"
+    "modules": "yarn clean",
+    "lint": "eslint --ext js,jsx,ts,tsx src",
+    "prettier": "prettier --write 'src/**/*.{js,jsx,ts,tsx,html,css,less,json}'"
   },
   "dependencies": {
-    "@spinnaker/core": "0.11.1",
-    "@spinnaker/pluginsdk": "0.0.32",
-    "@types/react-dom": "17.0.13",
-    "bufferutil": "4.0.1",
+    "@spinnaker/core": "0.16.0",
+    "@spinnaker/eslint-plugin": "1.0.13",
+    "@spinnaker/pluginsdk": "0.2.3",
+    "@spinnaker/pluginsdk-peerdeps": "0.0.18",
+    "@spinnaker/presentation": "0.1.0",
+    "@types/react-dom": "18.0.0",
+    "@uirouter/core": "6.0.4",
+    "@uirouter/react": "1.0.2",
     "graphql": "15.8.0",
-    "js-yaml": "3.13.1",
-    "lodash": "4.17.21",
+    "lodash-es": "4.17.15",
+    "prop-types": "15.6.1",
     "react": "16.14.0",
-    "react-ace": "6.4.0",
-    "react-dom": "16.14.0"
+    "react-dom": "16.14.0",
+    "rollup": "2.45.2",
+    "rxjs": "6.6.7"
   },
   "devDependencies": {
-    "@spinnaker/eslint-plugin": "1.0.7",
-    "@types/lodash": "4.14.64",
+    "@rollup/plugin-commonjs": "18.0.0",
+    "@rollup/plugin-json": "4.1.0",
+    "@rollup/plugin-node-resolve": "11.2.1",
+    "@rollup/plugin-replace": "2.4.2",
+    "@rollup/plugin-typescript": "8.2.1",
+    "@rollup/plugin-url": "6.0.0",
     "@types/react": "16.8.25",
-    "@typescript-eslint/eslint-plugin": "2.26.0",
-    "@typescript-eslint/parser": "2.26.0",
-    "eslint": "6.8.0",
+    "@typescript-eslint/eslint-plugin": "4.4.0",
+    "@typescript-eslint/parser": "4.4.0",
+    "bufferutil": "4.0.2",
+    "eslint": "7.10.0",
     "eslint-config-prettier": "6.12.0",
-    "eslint-plugin-react-hooks": "3.0.0",
-    "rollup": "2.30.0",
-    "shx": "0.3.2",
-    "typescript": "3.8.3"
+    "eslint-plugin-react-hooks": "4.1.2",
+    "npm-run-all": "4.1.5",
+    "prettier": "2.1.2",
+    "pretty-quick": "3.1.3",
+    "rollup-plugin-external-globals": "0.6.1",
+    "rollup-plugin-terser": "7.0.2",
+    "rollup-plugin-visualizer": "5.4.1",
+    "shx": "0.3.3",
+    "typescript": "4.7.0-dev.20220408",
+    "utf-8-validate": "5.0.3"
   },
   "files": [
     "build/dist"
